// src/app/model/cars-policy-car.ts
export interface CarsPolicyCar {
  brandId?: string;
  carBrandCode?: string;
  carBrandDesc?: string;
  carChassis?: string;
  carId?: string;
  carPlate?: string;
  carShapeId?: string;
  shape?: string;
  carTrademarkId?: string;
  carTrademarkCode?: string;
  carTrademarkDesc?: string;
  carYear?: string;
  policyNumber?: string;
  sysCreatedDate?: string;
  sysUpdatedDate?: string;
  sysUpdatedBy?: string;
  sysCreatedBy?: string;
}